requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==6.0.0
flake8-isort==6.0.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
Pillow==9.3.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.26.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
